# Special case single-qubit unitaries in _apply_unitary_to_tensor_