--- a/cirq/linalg/transformations.py
+++ b/cirq/linalg/transformations.py
@@ -149,7 +149,7 @@ def targeted_left_multiply(left_matrix: np.ndarray,
                      **({'out': out} if out is not None else {}))
 
 
-_TSliceAtom = Union[int, slice]
+_TSliceAtom = Union[int, slice, 'ellipsis']
 _TSlice = Union[_TSliceAtom, Sequence[_TSliceAtom]]