This is 20% faster than using einsum.

Setup:

```
import cirq
import numpy as np
c = cirq.Circuit.from_ops(cirq.X(q)**0.1 for q in cirq.LineQubit.range(24))
```

Before:

```
%time _ = c.apply_unitary_effect_to_state(dtype=np.complex64)
CPU times: user 4.05 s, sys: 611 ms, total: 4.66 s
Wall time: 4.66 s
```

After:

```
%time _ = c.apply_unitary_effect_to_state(dtype=np.complex64)
CPU times: user 3.38 s, sys: 279 ms, total: 3.66 s
Wall time: 3.66 s
```