@Strilanc 2018-10-18
For whatever reason, this approach is slower with 4x4 matrices than einsum. Goes from 1.95s to 3.38s in a test I did with 22 operations on 22 qubits. If I had to guess, I'd say it's because we do 12 slightly-faster passes over the data instead of one slightly-slower one. Whereas for a 2x2 matrix it's 2 slightly-faster passes vs one slightly-slower one, and all the "slightlys" line up to make that faster whereas 4x4 is slower.

(Qubit adjacency isn't relevant to this particular method.)